Download script for ECOMIP data
This is one example of a short-term NICAM run based on the EarthCARE satellite and the ORCESTRA field campaign (e.g., aircraft and ship observations).  The data are provided on the native grid with a horizontal resolution of 3.5 km and a temporal resolution of 30 minutes for both 2D and 3D variables for a short-term run.